Refactor UI module into smaller modules for better maintainability
